Split project details form to own component
